Move cross-feature use cases to core-domain to fix module isolation
