+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/di/DomainModule.kt
@@ -27,5 +27,9 @@ val domainModule = module {
         shareDomainModule,
         updateDomainModule,
         similarityDomainModule,
+        collectionsDomainModule,
+        galleryDomainModule,
+        promptsDomainModule,
+        searchDomainModule,
     )
 }
