+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/domain/usecase/AddExcludedTagUseCase.kt
@@ -1,0 +1,7 @@
+package com.riox432.civitdeck.domain.usecase
+
+import com.riox432.civitdeck.domain.repository.ExcludedTagRepository
+
+class AddExcludedTagUseCase(private val repository: ExcludedTagRepository) {
+    suspend operator fun invoke(tag: String) = repository.addExcludedTag(tag)
+}
