+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/domain/usecase/AutoSavePromptUseCase.kt
@@ -1,9 +1,9 @@
-package com.riox432.civitdeck.feature.prompts.domain.usecase
+package com.riox432.civitdeck.domain.usecase
 
 import com.riox432.civitdeck.domain.model.ImageGenerationMeta
 import com.riox432.civitdeck.domain.repository.SavedPromptRepository
 
-class SavePromptUseCase(private val repository: SavedPromptRepository) {
+class AutoSavePromptUseCase(private val repository: SavedPromptRepository) {
     suspend operator fun invoke(meta: ImageGenerationMeta, sourceImageUrl: String?) =
-        repository.save(meta, sourceImageUrl)
+        repository.autoSave(meta, sourceImageUrl)
 }
