+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/domain/usecase/ClearSearchHistoryUseCase.kt
@@ -1,0 +1,7 @@
+package com.riox432.civitdeck.domain.usecase
+
+import com.riox432.civitdeck.domain.repository.SearchHistoryRepository
+
+class ClearSearchHistoryUseCase(private val repository: SearchHistoryRepository) {
+    suspend operator fun invoke() = repository.clearAll()
+}
