+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/domain/usecase/GetExcludedTagsUseCase.kt
@@ -1,0 +1,7 @@
+package com.riox432.civitdeck.domain.usecase
+
+import com.riox432.civitdeck.domain.repository.ExcludedTagRepository
+
+class GetExcludedTagsUseCase(private val repository: ExcludedTagRepository) {
+    suspend operator fun invoke(): List<String> = repository.getExcludedTags()
+}
