+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/domain/usecase/RemoveExcludedTagUseCase.kt
@@ -1,0 +1,7 @@
+package com.riox432.civitdeck.domain.usecase
+
+import com.riox432.civitdeck.domain.repository.ExcludedTagRepository
+
+class RemoveExcludedTagUseCase(private val repository: ExcludedTagRepository) {
+    suspend operator fun invoke(tag: String) = repository.removeExcludedTag(tag)
+}
