+++ core/core-domain/src/commonMain/kotlin/com/riox432/civitdeck/domain/usecase/RemoveModelFromCollectionUseCase.kt
@@ -1,4 +1,4 @@
-package com.riox432.civitdeck.feature.collections.domain.usecase
+package com.riox432.civitdeck.domain.usecase
 
 import com.riox432.civitdeck.domain.repository.CollectionRepository
