+++ feature/feature-collections/src/commonMain/kotlin/com/riox432/civitdeck/feature/collections/presentation/CollectionsViewModel.kt
@@ -3,9 +3,9 @@ package com.riox432.civitdeck.feature.collections.presentation
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.riox432.civitdeck.domain.model.ModelCollection
-import com.riox432.civitdeck.feature.collections.domain.usecase.CreateCollectionUseCase
+import com.riox432.civitdeck.domain.usecase.CreateCollectionUseCase
+import com.riox432.civitdeck.domain.usecase.ObserveCollectionsUseCase
 import com.riox432.civitdeck.feature.collections.domain.usecase.DeleteCollectionUseCase
-import com.riox432.civitdeck.feature.collections.domain.usecase.ObserveCollectionsUseCase
 import com.riox432.civitdeck.feature.collections.domain.usecase.RenameCollectionUseCase
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.flow.SharingStarted
