+++ feature/feature-detail/build.gradle.kts
@@ -6,8 +6,6 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(libs.androidx.lifecycle.viewmodel)
-            implementation(project(":feature:feature-gallery"))
-            implementation(project(":feature:feature-collections"))
         }
     }
 }
