+++ feature/feature-gallery/build.gradle.kts
@@ -7,7 +7,6 @@ kotlin {
         commonMain.dependencies {
             implementation(libs.androidx.lifecycle.viewmodel)
             implementation(libs.kotlinx.serialization.json)
-            implementation(project(":feature:feature-prompts"))
         }
     }
 }
