+++ feature/feature-settings/build.gradle.kts
@@ -6,7 +6,6 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(libs.androidx.lifecycle.viewmodel)
-            implementation(project(":feature:feature-search"))
         }
     }
 }
