+++ feature/feature-settings/src/commonMain/kotlin/com/riox432/civitdeck/feature/settings/presentation/ContentFilterSettingsViewModel.kt
@@ -5,15 +5,15 @@ import androidx.lifecycle.viewModelScope
 import com.riox432.civitdeck.domain.model.HiddenModel
 import com.riox432.civitdeck.domain.model.NsfwBlurSettings
 import com.riox432.civitdeck.domain.model.NsfwFilterLevel
+import com.riox432.civitdeck.domain.usecase.AddExcludedTagUseCase
+import com.riox432.civitdeck.domain.usecase.GetExcludedTagsUseCase
 import com.riox432.civitdeck.domain.usecase.GetHiddenModelsUseCase
 import com.riox432.civitdeck.domain.usecase.ObserveNsfwBlurSettingsUseCase
 import com.riox432.civitdeck.domain.usecase.ObserveNsfwFilterUseCase
+import com.riox432.civitdeck.domain.usecase.RemoveExcludedTagUseCase
 import com.riox432.civitdeck.domain.usecase.SetNsfwBlurSettingsUseCase
 import com.riox432.civitdeck.domain.usecase.SetNsfwFilterUseCase
-import com.riox432.civitdeck.feature.search.domain.usecase.AddExcludedTagUseCase
-import com.riox432.civitdeck.feature.search.domain.usecase.GetExcludedTagsUseCase
-import com.riox432.civitdeck.feature.search.domain.usecase.RemoveExcludedTagUseCase
-import com.riox432.civitdeck.feature.search.domain.usecase.UnhideModelUseCase
+import com.riox432.civitdeck.domain.usecase.UnhideModelUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
